Use PDU interface
We only use it in a few places currently, enough to get things
to compile and run. We should be using it in much more places.

Similarly, in some places we cast []PDU back to []*Event, we need
to not do that. Likewise, in some places we cast PDU to *Event, we
need to not do that. For now though, hopefully this is a start.